pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maestria"
version = "0.1.0"
description = "Machine Learning project"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "flask",
    "requests",
    "python-dotenv",
    "scikit-learn",
    "gunicorn",
    "scikit-fuzzy",
    "networkx",
    "matplotlib",
    "mlxtend",
    "ace-tools"
]

[tool.hatch.build.targets.wheel]
packages = ["tema2"]

[tool.hatch.metadata]
allow-direct-references = true